// backend/src/common/enums/trip-status.enum.ts
export enum TripStatus {
  QUOTED = 'QUOTED',
  CONFIRMED = 'CONFIRMED',
  DRIVER_ASSIGNED = 'DRIVER_ASSIGNED',
  EN_ROUTE_TO_PICKUP = 'EN_ROUTE_TO_PICKUP',
  AT_PICKUP = 'AT_PICKUP',
  IN_TRANSIT = 'IN_TRANSIT',
  AT_STOP = 'AT_STOP',
  WAITING_AT_STOP = 'WAITING_AT_STOP',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  SOS_ACTIVE = 'SOS_ACTIVE',
}